Greatly limits how the dtype and device can be used in Tensor
Previously, setting `dtype` or `device` in the Tensor constructor
would incur a cast or copy operation respectively if the input data
were not already of the correct data type or on the right device.
These would be immediately evaluated, which could lead to extremely long tracing times.

Additionally, the implicit behavior is not in line with the rest of Tripy's
API design. With this change, users must explicitly call `cast` or `copy`.
This makes it clear *when* such an operation is actually happening.

+++ tripy/tests/integration/test_batchnorm.py
@@ -43,7 +43,7 @@ class TestBatchNorm:
         tp_batchnorm.running_var = tp.Tensor(batchnorm.running_var.detach())
 
         input = torch.randn(input_shape, dtype=torch_dtype).to("cuda")
-        tp_input = tp.Tensor(input, dtype=tp_dtype)
+        tp_input = tp.Tensor(input)
 
         output = eager_or_compiled(tp_batchnorm, tp_input)
