+++ tripy/tests/integration/test_groupnorm.py
@@ -48,7 +48,7 @@ def setup(dtype, input_shape, num_groups, num_channels):
     )
 
     input = torch.empty(*input_shape, dtype=torch_dtype, device="cuda").uniform_(0, 10)
-    tp_input = tp.Tensor(input, dtype=dtype)
+    tp_input = tp.Tensor(input)
     yield groupnorm, tp_groupnorm, tp_input
